requirements.txt
modelcontextprotocol~=0.1.0
fastapi~=0.115.12
uvicorn[standard]~=0.34.2
python-dotenv~=1.1.0
httpx~=0.28.1
requests~=2.32.3
mcp~=1.9.1
pydantic~=2.11.5
fastmcp~=2.5.0